create threejs test page
