+++ src/Assets/data/grassy_knoll_data.js
@@ -1,0 +1,434 @@
+// Matrix of color blocks including:
+  //   color,
+  //   duration of linear gradient animation
+  //   width of block
+  //   height of block
+
+  // DONE
+  const spring = [
+    [
+      {
+        color: "#C9DDD4",
+        duration: "12",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#C9DDD4",
+        duration: "10",
+        width: "100vw",
+        height: "7vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#C6D7CF",
+        duration: "15",
+        width: "100vw",
+        height: "9vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#BCCFC9",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B7CAC4",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B8C9C1",
+        duration: "20",
+        width: "100vw",
+        height: "8vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B7C7BD",
+        duration: "20",
+        width: "100vw",
+        height: "5vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B3C3B9",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#BBBEB3",
+        duration: "20",
+        width: "100vw",
+        height: "9vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B9BCAB",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+  ];
+
+  const summer = [
+    [
+      {
+        color: "#7CBED4",
+        duration: "12",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#74B7D2",
+        duration: "10",
+        width: "100vw",
+        height: "7vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#68AFCB",
+        duration: "15",
+        width: "100vw",
+        height: "9vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#65ACCA",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#61A6C5",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#63A8C7",
+        duration: "20",
+        width: "100vw",
+        height: "8vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#5DA2C1",
+        duration: "20",
+        width: "100vw",
+        height: "5vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#5D9EBC",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#6AA1BA",
+        duration: "20",
+        width: "100vw",
+        height: "9vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#78A6B3",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+  ];
+
+  const autumn = [
+    [
+      {
+        color: "E6E3EC",
+        duration: "12",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#E9E6EF",
+        duration: "10",
+        width: "100vw",
+        height: "7vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#EEEDF5",
+        duration: "15",
+        width: "100vw",
+        height: "9vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#ECEBF3",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#EEEDF3",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#F0EFF5",
+        duration: "20",
+        width: "100vw",
+        height: "8vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#F1F2F7",
+        duration: "20",
+        width: "100vw",
+        height: "5vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#F6F7F9",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#F2F3F7",
+        duration: "20",
+        width: "100vw",
+        height: "9vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#EDF8FB",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+  ];
+
+  const winter = [
+    [
+      {
+        color: "#849FBD",
+        duration: "12",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#99B2D0",
+        duration: "10",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#A6BCD4",
+        duration: "15",
+        width: "100vw",
+        height: "5vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B1C6D9",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#C4D2DB",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#D6DAD9",
+        duration: "20",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#E3E1D5",
+        duration: "20",
+        width: "100vw",
+        height: "5vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#F2E0C8",
+        duration: "20",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#F9D7B4",
+        duration: "20",
+        width: "100vw",
+        height: "4vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#F5C6A8",
+        duration: "20",
+        width: "100vw",
+        height: "6vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#E3AFAA",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#CFA9AA",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B5A8AF",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#A4A8B1",
+        duration: "20",
+        width: "100vw",
+        height: "3vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#B0B1B3",
+        duration: "17",
+        width: "100vw",
+        height: "1vh",
+        horizon: true,
+      },
+    ],
+    [
+      {
+        color: "#9FA0A2",
+        duration: "17",
+        width: "100vw",
+        height: "5vh",
+        horizon: true,
+      },
+    ]
+  ];
+
+  export {spring, summer, autumn, winter}
